+++ native/src/impl/Library.cc
@@ -3,6 +3,7 @@
 #include "../shaper/interop.hh"
 #include "../skottie/interop.hh"
 #include "../paragraph/interop.hh"
+#include "../svg/interop.hh"
 
 JNIEXPORT jint JNICALL JNI_OnLoad(JavaVM* vm, void* reserved) {
     JNIEnv* env;
@@ -20,6 +21,7 @@ extern "C" JNIEXPORT void JNICALL Java_org_jetbrains_skija_impl_Library__1nAfter
     skija::shaper::onLoad(env);
     skija::skottie::onLoad(env);
     skija::paragraph::onLoad(env);
+    skija::svg::onLoad(env);
 }
 
 JNIEXPORT void JNICALL JNI_OnUnload(JavaVM* vm, void* reserved) {
@@ -27,6 +29,7 @@ JNIEXPORT void JNICALL JNI_OnUnload(JavaVM* vm, void* reserved) {
     if (vm->GetEnv(reinterpret_cast<void**>(&env), JNI_VERSION_1_8) != JNI_OK)
         return;
 
+    skija::svg::onUnload(env);
     skija::paragraph::onUnload(env);
     skija::skottie::onUnload(env);
     skija::shaper::onUnload(env);