+++ shared/src/main/java/org/jetbrains/skija/svg/SVGLength.java
@@ -1,0 +1,19 @@
+package org.jetbrains.skija.svg;
+
+import lombok.*;
+import org.jetbrains.annotations.*;
+
+@AllArgsConstructor @Data @With
+public class SVGLength {
+    @ApiStatus.Internal public final float _value;
+    @ApiStatus.Internal public final SVGLengthUnit _unit;
+
+    @ApiStatus.Internal
+    public SVGLength(float value, int unit) {
+        this(value, SVGLengthUnit.values()[unit]);
+    }
+
+    public SVGLength(float value) {
+        this(value, SVGLengthUnit.NUMBER);
+    }
+}