+++ shared/src/main/java/org/jetbrains/skija/svg/SVGLengthType.java
@@ -1,0 +1,7 @@
+package org.jetbrains.skija.svg;
+
+public enum SVGLengthType {
+    HORIZONTAL,
+    VERTICAL,
+    OTHER;
+}