+++ shared/src/main/java/org/jetbrains/skija/svg/SVGLengthUnit.java
@@ -1,0 +1,15 @@
+package org.jetbrains.skija.svg;
+
+public enum SVGLengthUnit {
+    UNKNOWN,
+    NUMBER,
+    PERCENTAGE,
+    EMS,
+    EXS,
+    PX,
+    CM,
+    MM,
+    IN,
+    PT,
+    PC
+}