+++ shared/src/main/java/org/jetbrains/skija/svg/SVGNode.java
@@ -1,0 +1,27 @@
+package org.jetbrains.skija.svg;
+
+import java.lang.ref.*;
+import org.jetbrains.annotations.*;
+import org.jetbrains.skija.*;
+import org.jetbrains.skija.impl.*;
+
+public abstract class SVGNode extends RefCnt {
+    static { Library.staticLoad(); }
+
+    @ApiStatus.Internal
+    public SVGNode(long ptr) {
+        super(ptr);
+    }
+
+    @NotNull
+    public SVGTag getTag() {
+        try {
+            Stats.onNativeCall();
+            return SVGTag.values()[_nGetTag(_ptr)];
+        } finally {
+            Reference.reachabilityFence(this);
+        }
+    }
+
+    @ApiStatus.Internal public static native int _nGetTag(long ptr);
+}