+++ shared/src/main/java/org/jetbrains/skija/svg/SVGPreserveAspectRatioScale.java
@@ -1,0 +1,6 @@
+package org.jetbrains.skija.svg;
+
+public enum SVGPreserveAspectRatioScale {
+    MEET,
+    SLICE;
+}